Sunset intake module 🫡
