+++ .github/workflows/api-documentation.yml
@@ -55,13 +55,11 @@ jobs:
           export PYTHONPATH="${PYTHONPATH}:."
           python tools/api/generate-openapi.py rules_uu --module datarequest > build/api_datarequest.json
           python tools/api/generate-openapi.py rules_uu --module deposit > build/api_deposit.json
-          python tools/api/generate-openapi.py rules_uu --module intake > build/api_intake.json
 
       - name: Validate Yoda module API documentation
         run: |
           openapi-spec-validator build/api_datarequest.json
           openapi-spec-validator build/api_deposit.json
-          openapi-spec-validator build/api_intake.json
 
       - name: Deploy 🚀
         uses: JamesIves/github-pages-deploy-action@releases/v3
