+++ __init__.py
@@ -56,10 +56,6 @@ from integration_tests        import *
 # Import certain modules only when enabled.
 from .util.config import config
 
-if config.enable_intake:
-    from intake import *
-    from intake_vault import *
-
 if config.enable_datarequest:
     from datarequest import *
