--- tests/step_defs/api/test_api_intake.py
@@ -1,157 +1,0 @@
-# coding=utf-8
-"""Intake API feature tests."""
-
-__copyright__ = 'Copyright (c) 2020-2022, Utrecht University'
-__license__   = 'GPLv3, see LICENSE'
-
-import json
-
-from deepdiff import DeepDiff
-from pytest_bdd import (
-    given,
-    parsers,
-    scenarios,
-    then,
-)
-
-from conftest import api_request
-
-scenarios('../../features/api/api_intake.feature')
-
-
-@given('the Yoda intake list studies API is queried', target_fixture="api_response")
-def api_intake_list_studies(user):
-    return api_request(
-        user,
-        "intake_list_studies",
-        {}
-    )
-
-
-@given('the Yoda intake list datamanager studies API is queried', target_fixture="api_response")
-def api_intake_list_dm_studies(user):
-    return api_request(
-        user,
-        "intake_list_dm_studies",
-        {}
-    )
-
-
-@given(parsers.parse("the Yoda intake count total files API is queried with collection {collection}"), target_fixture="api_response")
-def api_intake_count_total_files(user, collection):
-    return api_request(
-        user,
-        "intake_count_total_files",
-        {"coll": collection}
-    )
-
-
-@given(parsers.parse("the Yoda intake list unrecognized files API is queried with collection {collection}"), target_fixture="api_response")
-def api_intake_list_unrecognized_files(user, collection):
-    return api_request(
-        user,
-        "intake_list_unrecognized_files",
-        {"coll": collection}
-    )
-
-
-@given(parsers.parse("the Yoda intake list datasets API is queried with collection {collection}"), target_fixture="api_response")
-def api_intake_list_datasets(user, collection):
-    return api_request(
-        user,
-        "intake_list_datasets",
-        {"coll": collection}
-    )
-
-
-@given(parsers.parse("the Yoda intake scan for datasets API is queried with collection {collection}"), target_fixture="api_response")
-def api_intake_scan_for_datasets(user, collection):
-    return api_request(
-        user,
-        "intake_scan_for_datasets",
-        {"coll": collection}
-    )
-
-
-@given(parsers.parse("the Yoda intake lock API is queried with dataset id {dataset_id} and collection {collection}"), target_fixture="api_response")
-def api_intake_lock_dataset(user, dataset_id, collection):
-    return api_request(
-        user,
-        "intake_lock_dataset",
-        {"path": collection, "dataset_ids": dataset_id.replace("*", "\n")}
-    )
-
-
-@given(parsers.parse("the Yoda intake unlock API is queried with dataset id {dataset_id} and collection {collection}"), target_fixture="api_response")
-def api_intake_unlock_dataset(user, dataset_id, collection):
-    return api_request(
-        user,
-        "intake_unlock_dataset",
-        {"path": collection, "dataset_ids": dataset_id.replace("*", "\n")}
-    )
-
-
-@given(parsers.parse("the Yoda intake dataset get details API is queried with dataset id {dataset_id} and collection {collection}"), target_fixture="api_response")
-def api_intake_dataset_get_details(user, dataset_id, collection):
-    return api_request(
-        user,
-        "intake_dataset_get_details",
-        {"coll": collection, "dataset_id": dataset_id.replace("*", "\t")}
-    )
-
-
-@given(parsers.parse("the Yoda intake dataset add comment API is queried with dataset id {dataset_id}, study id {study_id} and comment {comment}"), target_fixture="api_response")
-def api_intake_dataset_add_comment(user, dataset_id, study_id, comment):
-    return api_request(
-        user,
-        "intake_dataset_add_comment",
-        {"study_id": study_id, "dataset_id": dataset_id.replace("*", "\n"), "comment": comment}
-    )
-
-
-@given(parsers.parse("the Yoda intake report vault dataset counts per study API is queried with study id {study_id}"), target_fixture="api_response")
-def api_intake_report_vault_dataset_counts_per_study(user, study_id):
-    return api_request(
-        user,
-        "intake_report_vault_dataset_counts_per_study",
-        {"study_id": study_id}
-    )
-
-
-@given(parsers.parse("the Yoda intake report vault aggregated info API is queried with study id {study_id}"), target_fixture="api_response")
-def api_intake_report_vault_aggregated_info(user, study_id):
-    return api_request(
-        user,
-        "intake_report_vault_aggregated_info",
-        {"study_id": study_id}
-    )
-
-
-@given(parsers.parse("the Yoda intake report export study data API is queried with study id {study_id}"), target_fixture="api_response")
-def api_intake_report_export_study_data(user, study_id):
-    return api_request(
-        user,
-        "intake_report_export_study_data",
-        {"study_id": study_id}
-    )
-
-
-@then(parsers.parse("study {study} is returned"))
-def study_returned(api_response, study):
-    _, body = api_response
-
-    assert study in body['data']
-
-
-@then('debug')
-def debug(api_response):
-    _, body = api_response
-
-    assert 0, body
-
-
-@then(parsers.parse("the result is equivalent to {result}"))
-def result_equivalent_to(api_response, result):
-    _, body = api_response
-
-    assert DeepDiff(json.loads(result), body['data']) == {}
