+++ unit-tests/unit_tests.py
@@ -6,7 +6,6 @@ __license__   = 'GPLv3, see LICENSE'
 from unittest import makeSuite, TestSuite
 
 from test_group_import import GroupImportTest
-from test_intake import IntakeTest
 from test_policies import PoliciesTest
 from test_revisions import RevisionTest
 from test_schema_transformations import CorrectifyIsniTest, CorrectifyOrcidTest, CorrectifyScopusTest
@@ -21,7 +20,6 @@ def suite():
     test_suite.addTest(makeSuite(CorrectifyOrcidTest))
     test_suite.addTest(makeSuite(CorrectifyScopusTest))
     test_suite.addTest(makeSuite(GroupImportTest))
-    test_suite.addTest(makeSuite(IntakeTest))
     test_suite.addTest(makeSuite(PoliciesTest))
     test_suite.addTest(makeSuite(RevisionTest))
     test_suite.addTest(makeSuite(UtilMiscTest))
