+++ util/config.py
@@ -99,7 +99,6 @@ config = Config(environment=None,
                 enable_deposit=False,
                 enable_open_search=False,
                 enable_inactivity_notification=False,
-                enable_intake=False,
                 enable_datarequest=False,
                 enable_data_package_archive=False,
                 enable_data_package_download=False,
